✨ 새 체험 API 및 타입 정의 추가
